backend/requirements.txt
# Production Dependencies
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration & Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# HTTP Client (for external API calls if needed)
httpx==0.25.2

# LLM / AI
google-generativeai>=0.8.0
pyyaml>=6.0.1
